requirements.txt
pydantic==2.10.6
cachetools==5.5.1
diskcache
httpx
motor
cachetools
pyyaml
fastapi==0.111.0
uvicorn==0.30.1
google-cloud-documentai==2.29.0
google-api-core==2.19.0
python-multipart==0.0.9
vertexai==1.49.0
pdf2image==1.17.0
aiofiles==24.1.0
poppler-utils==0.1.0
python-dotenv==1.0.1